Backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
pydantic>=2.0.0
python-dotenv
transformers
torch